refactor & cleanup: Moved character.h and player.h from basepvp & reduced diff with ddnet
