+++ src/game/server/entities/character.h
@@ -88,7 +88,7 @@ public:
 	bool IncreaseHealth(int Amount);
 	bool IncreaseArmor(int Amount);
 
-	void GiveWeapon(int Weapon, bool Remove = false, int Ammo = -1);
+	void GiveWeapon(int Weapon, bool Remove = false);
 	void GiveNinja();
 	void RemoveNinja();
 	void SetEndlessHook(bool Enable);
