+++ src/game/server/instagib/includes/player.h
@@ -1,6 +1,6 @@
 // This file can be included several times.
 
-#include <game/server/gamemodes/base_pvp/player.h>
+#include <game/server/instagib/player.h>
 
 // Put your custom includes below this comment
 // to avoid merge conflicts when this file gets updated.
