+++ src/game/server/instagib/player.cpp
@@ -6,14 +6,12 @@
 #include <generated/protocol.h>
 
 #include <game/server/entities/character.h>
-#include <game/server/entity.h>
 #include <game/server/gamecontext.h>
 #include <game/server/gamecontroller.h>
 #include <game/server/instagib/sql_stats.h>
 #include <game/server/instagib/structs.h>
 #include <game/server/player.h>
 #include <game/server/score.h>
-#include <game/version.h>
 
 #include <optional>
